LayoutTests/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/srcdoc-in-sandbox-cannot-access-parent.html from frame with URL about:srcdoc. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 4: Uncaught TypeError: Cannot call method 'getElementById' of undefined
PASS
